requirements.txt
win11toast
gradio==4.14.0
gradio_client~=0.8.0
accelerate==0.25.0
transformers==4.36.2
diffusers[torch]==0.25.0
ftfy==6.1.1
# albumentations==1.3.0
opencv-python==4.7.0.68
einops==0.7.0
pytorch-lightning==1.9.0
bitsandbytes
tensorboard==2.10.1
lion-pytorch==0.0.6
prodigyopt==1.0
safetensors==0.4.2
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.20.1
# for BLIP captioning
# requests==2.28.2
# timm==0.6.12
# fairscale==0.4.13
# for WD14 captioning (tensorflow)
# tensorflow==2.10.1
# for WD14 captioning (onnx)
# onnx==1.14.1
# onnxruntime-gpu==1.16.0
# onnxruntime==1.16.0
# this is for onnx: 
# protobuf==3.20.3
# open clip for SDXL
# open-clip-torch==2.20.0
# For logging
rich==13.7.0
pyperclip
pillow~=10.2.0
#torch~=2.3.0+cu118
tqdm~=4.66.2
numpy~=1.26.4
#torchvision~=0.18.0+cu118
#xformers~=0.0.26.post1+cu118
packaging~=24.0
timm~=0.9.16
setuptools~=65.5.0
regex~=2023.12.25
# for Image utils
imagesize==1.4.1